make benchmark reader hartbeat inverval aware fix small memleak on tablereader destruction make timespread better configureable

+++ src/input/readers/Benchmark.h
@@ -45,7 +45,7 @@ private:
 	int add;
 	int stopspreadat;
 	double heartbeatstarttime;
-	int timedspread;
+	double timedspread;
 
 	string RandomString(const int len);
